Attach advanced search controller
