+++ app/views/searches/advanced.html.erb
@@ -1,6 +1,6 @@
 <% @title = t("advanced_search") %>
 
-<div class="home__search-wrap">
+<div class="home__search-wrap" data-controller="advanced-search">
   <%= form_tag search_url, id: "advanced-search", method: :get do %>
     <%= search_field_tag :query, params[:query], placeholder: t("layouts.application.header.search_gem_html"), autofocus: current_page?(root_url), id: "home_query", class: "home__search" %>
     <%= label_tag :home_query do %>
